Finished day 18 part 1

+++ src/2017/day_18/main/com/aoc2017/day18/Compiler.kt
@@ -43,7 +43,7 @@ object Compiler {
 
     private fun getValueHolder(str: String): ValueHolder {
         return if (str.matches(NUM_REGEX)) {
-            IntValue(str.toInt())
+            LongValue(str.toLong())
         } else {
             Register(str)
         }
